#small-refactoring-and-fix-directory-name-for-validation
deleted unused constructors
deleted empty lines
removed annotation folder
