+++ src/main/java/org/example/task/dto/BasicResponseDto.java
