+++ src/main/java/org/example/task/model/User.java
@@ -3,7 +3,6 @@ package org.example.task.model;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 import lombok.*;
-import org.example.task.validator.annotation.NameRestriction;
 import org.hibernate.proxy.HibernateProxy;
 
 import java.util.Objects;
@@ -20,10 +19,6 @@ public class User {
     private Long id;
     @NotBlank(message = "Username cannot be empty")
     private String username;
-
-    public User(String username) {
-        this.username = username;
-    }
 
     @Override
     public final boolean equals(Object o) {
